Implemented basic slider functionality with new watch-link api
