+++ index.html
@@ -89,6 +89,12 @@
         v-html="svgItem.svg">
       </div>
     </template>
+    <template id="user-sliders-template">
+      <div @vue:mounted="slider.updateElement()" v-for="slider in sliders.sliders" :id="slider.id" :class="slider.id">
+        <input type="range" :min="slider.min" :max="slider.max" :step="slider.step" :value="slider.value"
+          :onInput="e => slider.value = e.target.value">
+      </div>
+    </template>
     <div id="app" style="
             position: absolute;
             top: 2rem;
@@ -101,6 +107,7 @@
     ></div>
     <script class="user-imports"></script>
     <div class="user-svgs" style="position: absolute; top: 0; left: 0; z-index: 1" v-scope="UserSVGs()"></div>
+    <div class="user-sliders" style="position: absolute; top: 0; left: 0; z-index: 2" v-scope="UserSliders()"></div>
     <style>
       * {
         border: 0;
